# docs/swagger.yaml
basePath: /
definitions:
  bookmark.BookmarkRequest:
    properties:
      description:
        type: string
      url:
        example: https://www.google.com
        maxLength: 2048
        type: string
    required:
    - url
    type: object
  bookmark.DeleteResponse:
    properties:
      message:
        example: Success
        type: string
    type: object
  bookmark.GetBookmarksResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Bookmark'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/bookmark.Pagination'
    type: object
  bookmark.Pagination:
    properties:
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
    type: object
  bookmark.UpdateRequest:
    properties:
      description:
        example: Google
        maxLength: 255
        type: string
      id:
        type: string
      url:
        example: https://www.google.com
        maxLength: 2048
        type: string
    required:
    - id
    - url
    type: object
  bookmark.UpdateResponse:
    properties:
      message:
        example: Success
        type: string
    type: object
  dto.ShortenURLReq:
    properties:
      exp:
        description: Exp is the expiration time of the shortened URL in seconds.
        minimum: 5
        type: integer
      url:
        description: URL is the original URL to shorten.
        example: https://example.com
        type: string
    required:
    - exp
    - url
    type: object
  dto.ShortenURLRes:
    properties:
      code:
        description: Code is the shortened URL code returned to the client.
        example: AbCDeFK
        type: string
      message:
        description: Message provides a status message for the operation.
        example: Shorten URL generated successfully!
        type: string
    type: object
  model.Bookmark:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      updated_at:
        type: string
      url:
        type: string
      user_id:
        type: string
    type: object
  model.HealthReport:
    properties:
      dependency:
        additionalProperties:
          type: string
        example:
          ' "postgres"': ' "UP"}'
          '{"redis"': ' "UP"'
        type: object
      instance_id:
        example: cbe1a562-596b-45d0-bf8b-a999b23b184a
        type: string
      message:
        example: OK
        type: string
      service_name:
        example: bookmark_service
        type: string
    type: object
  model.User:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  response.ErrMessage:
    properties:
      details: {}
      message:
        type: string
    type: object
  user.LoginRequest:
    properties:
      password:
        example: Password123@
        type: string
      username:
        example: johndoe
        maxLength: 20
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  user.LoginResponse:
    properties:
      data:
        example: jwt_token
        type: string
      message:
        example: Logged in successfully!
        type: string
    type: object
  user.RegisterData:
    properties:
      created_at:
        example: "2026-07-21T12:34:56Z"
        type: string
      display_name:
        example: John Doe
        type: string
      email:
        example: john.doe@example.com
        type: string
      id:
        example: 7d80f755-7dce-4c95-b8bf-75bb8e240ef2
        type: string
      updated_at:
        example: "2026-07-21T12:34:56Z"
        type: string
      username:
        example: johndoe
        type: string
    type: object
  user.RegisterRequest:
    properties:
      display_name:
        example: John Doe
        maxLength: 20
        minLength: 3
        type: string
      email:
        example: john.doe@example.com
        type: string
      password:
        example: Password123@
        type: string
      username:
        example: johndoe
        maxLength: 20
        minLength: 3
        type: string
    required:
    - display_name
    - email
    - password
    - username
    type: object
  user.RegisterResponse:
    properties:
      data:
        $ref: '#/definitions/user.RegisterData'
      message:
        type: string
    type: object
  user.UpdateInfoRes:
    properties:
      message:
        example: Edit current user successfully!
        type: string
    type: object
  user.UpdateUserInfoReq:
    properties:
      display_name:
        type: string
      email:
        type: string
    required:
    - display_name
    - email
    type: object
info:
  contact: {}
  description: This is a simple REST API for a bookmark service
  title: Bookmark Service API
  version: "2.0"
paths:
  /health-check:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HealthReport'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/model.HealthReport'
      summary: Health Check
      tags:
      - Health Check
  /v1/bookmarks:
    delete:
      description: Deletes a bookmark belonging to the authenticated user by its ID.
      parameters:
      - description: Bookmark ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Bookmark deleted successfully
          schema:
            $ref: '#/definitions/bookmark.DeleteResponse'
        "400":
          description: Bad Request / Invalid input
          schema:
            $ref: '#/definitions/response.ErrMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrMessage'
        "404":
          description: Bookmark not found
          schema:
            $ref: '#/definitions/response.ErrMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrMessage'
      security:
      - BearerAuth: []
      summary: Delete bookmark
      tags:
      - Bookmarks
    get:
      description: Retrieves all bookmarks belonging to the authenticated user with
        pagination.
      parameters:
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Bookmarks retrieved successfully
          schema:
            $ref: '#/definitions/bookmark.GetBookmarksResponse'
        "400":
          description: Bad Request / Invalid input
          schema:
            $ref: '#/definitions/response.ErrMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrMessage'
      security:
      - BearerAuth: []
      summary: Get user bookmarks
      tags:
      - Bookmarks
    post:
      consumes:
      - application/json
      description: Creates a new bookmark for the authenticated user with the provided
        description and URL.
      parameters:
      - description: Bookmark request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/bookmark.BookmarkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Bookmark created successfully
          schema:
            $ref: '#/definitions/model.Bookmark'
        "400":
          description: Bad Request / Invalid input
          schema:
            $ref: '#/definitions/response.ErrMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrMessage'
        "409":
          description: Conflict / Duplicate bookmark code
          schema:
            $ref: '#/definitions/response.ErrMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrMessage'
      security:
      - BearerAuth: []
      summary: Add Bookmark
      tags:
      - Bookmarks
    put:
      consumes:
      - application/json
      description: Updates the description and/or URL of an existing bookmark belonging
        to the authenticated user.
      parameters:
      - description: Bookmark ID
        in: query
        name: id
        required: true
        type: string
      - description: Update bookmark request payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/bookmark.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Bookmark updated successfully
          schema:
            $ref: '#/definitions/bookmark.UpdateResponse'
        "400":
          description: Bad Request / Invalid input
          schema:
            $ref: '#/definitions/response.ErrMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrMessage'
      security:
      - BearerAuth: []
      summary: Update bookmark
      tags:
      - Bookmarks
  /v1/links/redirect/{code}:
    get:
      description: Resolves a short code and redirects the client to the original
        URL.
      parameters:
      - description: Short code
        in: path
        name: code
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "302":
          description: Redirect to the original URL
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Code not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Redirect to original URL
      tags:
      - ShortenURL
  /v1/links/shorten:
    post:
      consumes:
      - application/json
      description: Creates a unique short code for the given URL and returns it.
      parameters:
      - description: Shorten URL request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ShortenURLReq'
      produces:
      - application/json
      responses:
        "200":
          description: Shorten URL response
          schema:
            $ref: '#/definitions/dto.ShortenURLRes'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create Shorten Link
      tags:
      - ShortenURL
  /v1/self/info:
    get:
      consumes:
      - application/json
      description: get your current information
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            properties:
              data:
                $ref: '#/definitions/model.User'
            type: object
      security:
      - BearerAuth: []
      summary: get your current information
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Updates the authenticated user's display name and email address.
      parameters:
      - description: Update user info request payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserInfoReq'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            $ref: '#/definitions/user.UpdateInfoRes'
        "400":
          description: Invalid request input or validation details
          schema:
            $ref: '#/definitions/response.ErrMessage'
        "401":
          description: Unauthorized access token
          schema:
            $ref: '#/definitions/response.ErrMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrMessage'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - User
  /v1/users/login:
    post:
      consumes:
      - application/json
      description: Logs in a user with the provided username and password, returning
        a JWT token.
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login response
          schema:
            $ref: '#/definitions/user.LoginResponse'
        "400":
          description: Bad Request / Invalid credentials
          schema:
            $ref: '#/definitions/response.ErrMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrMessage'
      summary: User Login
      tags:
      - User
  /v1/users/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with the provided username, display
        name, password, and email.
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Register response
          schema:
            $ref: '#/definitions/user.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"